+++ R-normalization.R
@@ -1,5 +1,3 @@
-library(edgeR)
-library(DESeq2)
 
 normalized_tmm_data <- function(input_data) {
   #'
@@ -8,9 +6,9 @@ normalized_tmm_data <- function(input_data) {
   #'
   #' @param input_data The input data passed into the function
   #
-  dgList <- DGEList(counts=input_data)
-  dgList <- calcNormFactors(dgList, method="TMM")
-  edger_tmm_normalized_data <- cpm(dgList, normalized.lib.sizes=TRUE) # Obtain the absolute TMM values of the data
+  dgList <- edgeR::DGEList(counts=input_data)
+  dgList <- edgeR::calcNormFactors(dgList, method="TMM")
+  edger_tmm_normalized_data <- edgeR::cpm(dgList, normalized.lib.sizes=TRUE) # Obtain the absolute TMM values of the data
 
   return(edger_tmm_normalized_data)
 }
@@ -27,9 +25,9 @@ normalized_deseq2_data <- function(input_data) {
   meta_data <- data.frame(matrix(NA, nrow = ncol(input_data), ncol = 1))
   rownames(meta_data) <- colnames(input_data)
 
-  dds <- DESeqDataSetFromMatrix(countData = round(input_data), colData = meta_data, design=~1)
-  dds <- estimateSizeFactors(dds)
-  deseq2_normalized_data <- counts(dds, normalized=TRUE)
+  dds <- DESeq2::DESeqDataSetFromMatrix(countData = round(input_data), colData = meta_data, design=~1)
+  dds <- DESeq2::estimateSizeFactors(dds)
+  deseq2_normalized_data <- DESeq2::counts(dds, normalized=TRUE)
 
   return(deseq2_normalized_data)
 }
@@ -42,7 +40,7 @@ normalized_tpm_data <- function(input_data, gene_length) {
   #' @param input_data The input data passed into the function
   #
 
-  x <- input_data / gene_length
+  x <- input_data / as.numeric(gene_length)
   tpm <- t( t(x) * 1e6 / colSums(x) )
 
   return(tpm)