+++ wrangling.py
@@ -4,25 +4,27 @@
    supplementary files into h5ad for input into DHART. 
 """
 
-from ast import arg
 import string
 import sys
 import logging
 import argparse
 import re
-from importlib_metadata import metadata
+import json
+import tarfile
+
+import numpy as np
 import pandas as pd
-import rpy2.robjects as robjects
-from rpy2.robjects import pandas2ri
-from rpy2.robjects.conversion import localconverter
-from rpy2.robjects.packages import importr
-import scipy as sc
+from scipy.sparse import csr_matrix
 import anndata as ad
+
+import rpy2.robjects as robjects
+from rpy2.robjects import pandas2ri, numpy2ri
+from rpy2.robjects.conversion import localconverter
+# from rpy2.robjects.packages import importr
+
 import utils.utils as utils
 import utils.data_utils as data_utils
-import mygene
 
-import tarfile
 from pathlib import Path
 from paths import PROJECT_ROOT
 UTILS_PATH = PROJECT_ROOT / 'utils'
@@ -30,108 +32,116 @@ UTILS_PATH = PROJECT_ROOT / 'utils'
 logger = logging.getLogger(__name__)
 
 
-required_files = []
+# defined as in accession.py
+data_choices = ['single-cell RNA-Seq', 
+                'bulk RNA-Seq', 
+                'microarray', 
+                'ChIP-Seq', 
+                'ATAC-Seq']
+
 
 def main():
     parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter,
                                      description="""Data wrangling wrapper.""")
 
-    parser.add_argument('-d', '--dest', help="""Path to directory where GEO files reside.""",
-                        type=Path, default=Path(__file__).parent)
+    parser.add_argument('-d', '--dest', help="""Path to directory where GEO files reside, 
+                        including output from accession script and/or pre-processed bulk 
+                        RNA-seq count matrix.""", type=Path, default=Path(__file__).parent)
     
     parser.add_argument('-n', '--name', help="""Name of H5AD output.""", type=str, 
                         default='adata')
 
-    parser.add_argument('-b', '--bulk', help="""Specifies whether input file is bulk or single-cell data.""", 
-                        action='store_true')
-
-    parser.add_argument('--species', help="""Specifies the species to be used for missing gene ID or gene name info in bulk data. 
-                        If both gene ID and name are provided, this is silently ignored. 
-                        This is also silently ignored, if the -b flag isn't set. The default value is human. """, 
-                        type=str, default='human')
-
-    parser.add_argument('--gene-info', help="""Indicate presence of gene ID and/or gene name.""", type=str, 
-                        default='ID', choices=['ID', 'name', 'ID+name'])
-
-    parser.add_argument('--gene-info-id', help="""Provide input gene types as defined by mygene. 
-                        The default value is ensembl.gene, hence ensembl is assumed as the standard input format.""", 
-                        type=str, default='ensembl.gene')
-    
     parser.add_argument('-fmt', '--input-format', help="""Input format: either MEX (e.g. TSV
                         and MTX), or TXT (e.g. TXT, TSV, or CSV). Market Exchange (MEX) format
                         is for gene-barcode matrix output from Cell Ranger; TXT format is 
-                        for read count matrices (single file, genes x barcodes/wells).""", 
-                        type=str, choices=['MEX', 'TXT']) # This may be reintroduced for scRNA processing only: , required=True
+                        for read count matrices (single file, genes x barcodes/wells/samples).""", 
+                        type=str, choices=['MEX', 'TXT'], default=None)
     
     parser.add_argument('--mex-gene', help="""File name pattern, typically "features" or
                         "genes", used with [--input-format MEX]. Silently ignored if
-                        [--input-format TXT].""", default='feature', type=str)
+                        [--input-format TXT].""", type=str, default='feature')
     
     parser.add_argument('--mex-barcode', help="""File name pattern, typically "barcodes",
                         used with [--input-format MEX]. Silently ignored if
-                        [--input-format TXT].""", default='barcode', type=str)
+                        [--input-format TXT].""", type=str, default='barcode')
     
     parser.add_argument('--mex-matrix', help="""File name pattern, typically "matrix",
                         used with [--input-format MEX]. Silently ignored if
-                        [--input-format TXT].""", default='matrix', type=str)
+                        [--input-format TXT].""", type=str, default='matrix')
 
     parser.add_argument('--mex-gene-ncols', help="""Number of feature file columns, 
                         used with [--input-format MEX]. If [--mex-gene-ncols 1], it should
                         be "gene_symbol", otherwise they must be in the following 
                         order: ["gene_id", "gene_symbol", "feature_type"].
-                        Silently ignored if [--input-format TXT].""", choices=[1, 2, 3], 
-                        default=2, type=int)
-                        
-    parser.add_argument('--mex-gene-var', help="""Column to use as "var_names". If
-                        [--mex-gene-ncols 1], uses whatever is there. Silently ignored if
-                        [--input-format TXT].""", choices=['gene_symbol', 'gene_id'], 
-                        default='gene_symbol', type=str)
+                        Silently ignored if [--input-format TXT].""", type=int, 
+                        choices=[1, 2, 3], default=2)
     
     parser.add_argument('--txt-ext', help="""Extension that indicates the file type. 
                         If None, uses extension of filename, see the anndata read_txt 
                         function. Silently ignored if [--input-format MEX].""", 
-                        default=None, type=str)
+                        type=str, default=None)
     
     parser.add_argument('--txt-delimiter', help="""If None, will split at arbitrary 
                         number of white spaces, see the anndata read_txt function.
-                        Silently ignored if [--input-format MEX].""", default=None,
-                        type=str)
+                        Silently ignored if [--input-format MEX].""", type=str,
+                        default=None)
     
     parser.add_argument('--txt-first-column-names', help="""Assume the first column 
                         stores row names. This is only necessary if these are not 
                         strings: strings in the first column are automatically assumed 
                         to be row names. See anndata read_txt function.
-                        Silently ignored if [--input-format MEX].""", default=False,
-                        type=bool)
+                        Silently ignored if [--input-format MEX].""", 
+                        action='store_true')
     
     parser.add_argument('--txt-first-row-names', help="""Assume the first row stores 
                         column names (barcodes/wells). The behaviour of this flag is 
-                        mostly dependent on the anndata read_txt function. If the 
-                        first line starts with "#", this is ignored. If the first line 
-                        does not start with "#", it might wrongly be read into X 
-                        (and var), e.g. if column names are well numbers. If True, 
-                        then column names are read separately and re-assigned, and this 
-                        depends on [--txt-delimiter]. If False, only the first row is 
-                        removed, and obs_names are numbered by default. Silently ignored 
-                        if [--input-format MEX].""", default=False, type=bool)
+                        mostly dependent on the anndata read_txt function. In general, 
+                        it might wrongly be read into X (and var), e.g. if column names 
+                        are well numbers. If True, then column names are read separately 
+                        and re-assigned, and this depends on [--txt-delimiter]. If False, 
+                        fall back entirely to read_txt.. Silently ignored 
+                        if [--input-format MEX].""", action='store_true')
+    
+    parser.add_argument('--txt-gene-cols', help="""Content of first column(s), 
+                        used with [--input-format TXT]. [--txt-gene-cols] can either be
+                        "gene_id", "gene_symbol", or both, but then they must be in the 
+                        following order: ["gene_id", "gene_symbol"]. Any additional column
+                        will cause an error. Silently ignored if [--input-format MEX].""", 
+                        type=str, choices=["gene_id", "gene_symbol", "both"], 
+                        default="gene_id")
+    
+    parser.add_argument('--gene-var', help="""Column to use as "var_names". Can be used 
+                        with either [--input-format] options.""", type=str, 
+                        choices=['gene_symbol', 'gene_id'], default='gene_symbol')
     
     parser.add_argument('-p', '--pattern', help="""A space separated list of patterns 
                         used to glob files. Can be used with either [--input-format] options.
                         Search is performed independently (and before) file name pattern 
-                        search for [--input-format MEX].""", default='', type=str, nargs='*')
+                        search for [--input-format MEX].""", type=str, default='', nargs='*')
     
     parser.add_argument('--normalised', help="""If this flag is present, then 
-                        files are treated as normalised input.""", action='store_true')
-    # TODO: if input is normalised, what do we do?
+                        files are treated as normalised input. This has currently no effect:
+                        for single-cell (presumably TXT fmt) wrangling is the same, for bulk,
+                        this overrides [--do-normalise], if set.""", action='store_true')
     
     parser.add_argument('--do-normalise', help="""If this flag is present, then 
-                        output is also written as normalised, in addition to raw.
+                        output is written as normalised (only for bulk, otherwise ignored). 
                         Input files must be raw.""", action='store_true')
 
-    parser.add_argument('--normalization-method', help="""Set the normalization method. 
-                        Flag is silently ignored if --do-normalize isn't set""", 
-                        default='edgeR-TMM', choices=['edgeR-TMM', 'deseq2', 'TPM'])
+    parser.add_argument('--normalization-method', help="""Set the normalization method
+                        for bulk RNA-seq only.""", required='--do-normalise' in sys.argv, 
+                        type=str, choices=['TMM', 'DESeq', 'TPM'], default='TMM')
     
+    parser.add_argument('--gene-lengths', help="""If [--do-normalise] is set, and 
+                        [--normalization-method] is TPM, gene lengths required for 
+                        normalization. A single column (one length per row), wtihout 
+                        header, where genes are in the same order as the read count
+                        matrix.""", type=str, default=None)
+    
+    parser.add_argument('--mygene-scopes', help="""Type of types of identifiers as defined 
+                        by MyGene.py. Refer to official MyGene.info docs for full list of 
+                        fields.""", type=str, default='ensembl.gene')
+        
     utils.add_logging_options(parser)
     args = parser.parse_args(args=None if sys.argv[1:] else ['--help'])
     utils.update_logging(args)
@@ -139,37 +149,48 @@ def main():
     msg = f'[wrangling]: {" ".join(sys.argv)}'
     logger.info(msg)
     
-    # Set required files according to what data was provided
-    if args.bulk: 
-        required_files = ['bulk_input.csv', 'observations.tab.gz']
-    else:
-        required_files = ['_filelist.txt', '_RAW.tar', 'observations.tab.gz']
-        
-    # first check that all required files are available
-    all_files = [p.name for p in list(filter(Path.is_file, args.dest.glob('*.*')))]
-    pattern = fr'.*({"|".join(required_files)})'
-    filenames = list(data_utils.glob_re(pattern, all_files))
-    if len(filenames) != len(required_files):
-        msg = f'Missing input files! Required files are: {", ".join(required_files)}. ' \
-              f'Only these files were found: {", ".join(filenames)}'
+    # load json config and observations dataframe (from accession.py)
+    fl = list(filter(Path.is_file, args.dest.glob('*.json')))
+    if len(fl) != 1:
+        msg = "[--dest] must include a single configuration file (*.json). " \
+              f"{len(fl)} json file(s) were found! Check accession.py logs. " \
+              "Remove non-essential files or specify a different directory. " \
+              "Terminating program execution!"
         logger.critical(msg)
         return
+    with open(fl[0], "r") as f:
+        config = json.load(f)
 
-    if not args.bulk:
-        # This section applies to scRNA data
-
-        # must be consistent with the order of required_files
-        filen = [f for f in filenames if required_files[0] in f][0]
+    filen = Path(args.dest, "observations.tab.gz")
+    if not filen.is_file():
+        msg = "Missing input file observations.tab.gz under [--dest]. " \
+              "Check accession.py logs or specify a different directory. " \
+              "Terminating program execution!"
+        logger.critical(msg)
+        return
+    observations = pd.read_csv(filen, sep="\t")
+    
+    # output directory
+    h5ad_dir = Path(args.dest, 'h5ad')
+    h5ad_dir.mkdir(parents=True, exist_ok=False)
+        
+    msg = f"Processing {config['dataset_type']} {config['geo_accession']}. "
+    logger.info(msg)
+    
+    # data wrangling specific to single-cell 
+    if "single" in config['dataset_type'].lower():
+        filen = Path(args.dest, f"{config['geo_accession']}_filelist.txt")
+        if not filen.is_file():
+            msg = f"Missing input file {config['geo_accession']}_filelist.txt under [--dest]. " \
+                  "Terminating program execution!"
+            logger.critical(msg)
+            return
+        
         filelist = pd.read_csv(
-            Path(args.dest, filen),
+            filen,
             header=None,
             names=['File', 'Name', 'Time', 'Size', 'Type'],
             comment='#',
-            sep='\t'
-        )
-        filen = [f for f in filenames if required_files[2] in f][0]
-        observations = pd.read_csv(
-            Path(args.dest, filen),
             sep='\t'
         )
         tarname = Path(args.dest, filelist[filelist.File=='Archive'].Name[0])
@@ -199,16 +220,25 @@ def main():
                 'mex_barcode': args.mex_barcode,
                 'mex_matrix': args.mex_matrix,
                 'n_vars': args.mex_gene_ncols,
-                'var_names': args.mex_gene_var
+                'var_names': args.gene_var,
+                'feature': 'gene_symbol',
             }
-        else: # TXT
+        elif args.input_format == 'TXT':
             parse_fmt = data_utils.parse_txt_fmt
             kwargs = {
                 'ext': args.txt_ext,
                 'delimiter': args.txt_delimiter,
                 'first_column_names': args.txt_first_column_names,
-                'first_row_names': args.txt_first_row_names
+                'first_row_names': args.txt_first_row_names,
+                'n_vars': 2 if args.txt_gene_cols == 'both' else 1,
+                'var_names': args.gene_var,
+                'feature': args.txt_gene_cols,
             }
+        else:
+            msg = "[--input-format] undefined! Cannot process files." \
+                  "Terminating program execution!"
+            logger.critical(msg)
+            return
                 
         # extract
         extract_dir = Path(args.dest, 'extract')
@@ -255,263 +285,149 @@ def main():
                 logger.warning(msg)
                 
         # files are ready to be read in...
-        h5ad_dir = Path(args.dest, 'h5ad')
-        h5ad_dir.mkdir(parents=True, exist_ok=False)
-        if not args.normalised:
-            raw = parse_fmt(
-                extract_dir,
-                observations,
-                filelist,
-                **kwargs
-            )
-            adata = ad.concat(
-                raw, 
-                join='outer', 
-                index_unique="-", 
-                label='batch', 
-                merge="same")
-            adata.write_h5ad(Path(h5ad_dir, f'{args.name}_raw.h5ad'))
-            
-            if args.do_normalise:
-                pass
-            
-
-        else: # data is already normalised
-            pass
-    else: 
-        # This section applies to bulk RNA data
-
-        # must be consistent with the order of required_files
-        filen = [f for f in filenames if required_files[0] in f][0]
-        bulkdata = pd.read_csv(
-            Path(args.dest, filen),
-            header=0,
-            comment='#',
+        # currently same processing whether data is raw or normalised (this shouln't 
+        # really occur, given input formats)
+        # if not args.normalised:
+        # else: # data is already normalised
+        #    pass
+        
+        h5ad_name = f"{args.name}_raw.h5ad"
+        raw = parse_fmt(
+            extract_dir,
+            observations,
+            filelist,
+            **kwargs
         )
-        filen = [f for f in filenames if required_files[1] in f][0]
-        observations = pd.read_csv(
-            Path(args.dest, filen),
-            sep='\t'
+        adata = ad.concat(
+            raw, 
+            join='outer', 
+            index_unique="-", 
+            label='batch', 
+            merge="same"
         )
-        
-        # Read data from bulk_input.csv
-        if args.gene_info == 'ID':
-            # Only the gene ID is present
-            logger.debug('gene ID is present')
-
-            gene_metadata = bulkdata.iloc[:,[0]]
-            gene_metadata_series = bulkdata.iloc[:,0]
-            sampledata = bulkdata.iloc[:,1:]
-            print("input data shape: " + str(sampledata.shape))
-
-            gene_format = args.gene_info_id
-
-            mg = mygene.MyGeneInfo()
-            gene_metadata_list = gene_metadata_series.tolist()
-            species_from_args = args.species
-
-            # Run the mygene query
-            mygene_query_result = mg.querymany(gene_metadata_list, scopes=gene_format, fields='symbol,'+gene_format, species=species_from_args, as_dataframe=True)
-
-            # Extract just the symbols from the query result
-            matching_table = mygene_query_result["symbol"]
-
-            # Map the symbols to the existing metadata table
-            gene_metadata["name"] = gene_metadata["ID"].map(matching_table)
             
-            logger.debug(gene_metadata)
-            logger.debug('gene_metadata shape: ' + str(gene_metadata.shape))
-
-
-        if args.gene_info == 'name': 
-            # Only the gene name is present
-            logger.debug('gene name is present')
-
-            gene_metadata = bulkdata.iloc[:,[0]]
-            gene_metadata_series = bulkdata.iloc[:,0]
-            sampledata = bulkdata.iloc[:,1:]
-            logger.debug('input data shape: ' + str(sampledata.shape))
-
-            mg = mygene.MyGeneInfo()
-            gene_metadata_list = gene_metadata_series.tolist()
-            species_from_args = args.species
-
-            mygene_query_result = mg.querymany(gene_metadata_list, scopes='symbol', fields='ensembl.gene', species=species_from_args, as_dataframe=True)
-
-            matching_table = mygene_query_result[["ensembl.gene"]]
-
-            # Move symbols to their own column
-            matching_table = matching_table.reset_index(level=0)
-            
-            matching_table = matching_table.rename(columns={'ensembl.gene': 'ID'})
-
-            gene_metadata = (gene_metadata.merge(matching_table, left_on='name', right_on='query').reindex(columns=['ID', 'name']))
-
-            # Map the symbols to the existing metadata table
-            # This could be used to drop duplicates, however different IDs for symbols would then be lost. Merge above is preferred for now. 
-            #matching_table = matching_table.rename(columns={'query': 'name'})
-            #gene_metadata["ID"] = gene_metadata["name"].map(matching_table.drop_duplicates('name'))
-
-            logger.debug(gene_metadata)
-            logger.debug('gene_metadata shape: ' + str(gene_metadata.shape))
-
-        if args.gene_info == 'ID+name':
-            # Both the gene ID and name are present
-            logger.debug('gene ID and name are present')
-
-            gene_metadata = bulkdata.iloc[:,0:2]
-            sampledata = bulkdata.iloc[:,2:]
-            print("input data shape: " + str(sampledata.shape))
-
-        else: 
-            # Unrecognized gene_info parameter
-            msg = f'Unrecognized gene_info parameter. '
-            logger.critical(msg)
-        
-        
-
         if args.do_normalise:
-            # Data needs to be normalized, check for normalization method and normalize accordingly
-            # Defining the R script and loading the instance in Python
+            msg = "[--do-normalise] is set, but this option has currently no " \
+                  "effect on single-cell RNA-seq. Continuing!"
+            logger.warning(msg)
+        
+        
+    elif "bulk" in config['dataset_type'].lower():
+        # force text format - pre-formatted input
+        args.input_format = "TXT"
+        h5ad_name = f"{args.name}_normalized.h5ad"
+        if args.normalised:
+            args.do_normalise = False
+            msg = "Input treated as normalized data. Forcing [--do-normalise] to False." \
+                  "Continuing!"
+            logger.warning(msg)
+        if not args.normalised and not args.do_normalise:
+            h5ad_name = f"{args.name}_raw.h5ad"
+            msg = "Input treated as un-normalized data, but [--do-normalise] is unset." \
+                  "Continuing, but bulk RNA-seq data will NOT be normalized!"
+            logger.warning(msg)
+        all_files = [p.name for p in list(filter(Path.is_file, args.dest.glob('*.*')))]
+        # any TXT format, compressed or not
+        filenames = list(data_utils.glob_re(f"{config['geo_accession']}_read_count", all_files))
+        if len(filenames) != 1:
+            msg = "[--dest] must include a pre-formatted read count matrix  " \
+                  f"with the following name: {config['geo_accession']}_read_count. " \
+                  "Terminating program execution!"
+            logger.critical(msg)
+            return
+        filelist = pd.DataFrame({"Sample": [config['geo_accession']], "Name": filenames})
+        parse_fmt = data_utils.parse_txt_fmt
+        kwargs = {
+            'ext': args.txt_ext,
+            'delimiter': args.txt_delimiter,
+            'first_column_names': args.txt_first_column_names,
+            'first_row_names': args.txt_first_row_names,
+            'n_vars': 2 if args.txt_gene_cols == 'both' else 1,
+            'var_names': args.gene_var,
+            'feature': args.txt_gene_cols,
+            'bulk': True
+        }
+        adata = parse_fmt(
+            args.dest,
+            observations,
+            filelist,
+            **kwargs
+        )[config['geo_accession']]
+        
+        # bulk RNA-seq-specific normalization
+        if args.do_normalise:
+            msg = f"Normalizing using {args.normalization_method}."
+            logger.info(msg)
+            
+            # defining R script and loading instance in Python
             r = robjects.r
             rootpath = Path(__file__).parent
             filepath = str(Path(rootpath, 'R-normalization.R'))
-            print(filepath)
-
-            if args.normalization_method == 'edgeR-TMM':
-                # edgeR's TMM normalization selected
-                r['source'](filepath)
-                
-                # Load the function defined in R.
-                edgeR_TMM = robjects.globalenv['normalized_tmm_data']
-
-                # Convert it into an R object to pass into the R function
-                with localconverter(robjects.default_converter + pandas2ri.converter):
-                    input_data_r = robjects.conversion.py2rpy(sampledata)
-                
-                #Invoke the R function and get the result
-                df_result_r = edgeR_TMM(input_data_r)
-                
-                # Convert it back to a pandas dataframe.
-                with localconverter(robjects.default_converter + pandas2ri.converter):
-                    df_result = robjects.conversion.rpy2py(df_result_r)
-                
-                logger.debug(df_result)
-                print(df_result)
-                print(df_result.shape)
-
-            if args.normalization_method == 'deseq2':
-                # deseq2 normalization selected
-
-                # Defining the R script and loading the instance in Python
-                r = robjects.r
-                r['source'](filepath)
-
-                # Load the function defined in R.
-                deseq2 = robjects.globalenv['normalized_deseq2_data']
-
-                # Convert it into an R object to pass into the R function
-                with localconverter(robjects.default_converter + pandas2ri.converter):
-                    input_data_r = robjects.conversion.py2rpy(sampledata)
-
-                #Invoke the R function and get the result
-                df_result_r = deseq2(input_data_r)
-
-                # Convert it back to a pandas dataframe.
-                with localconverter(robjects.default_converter + pandas2ri.converter):
-                    df_result = robjects.conversion.rpy2py(df_result_r)
-
-                logger.debug(df_result)
-                print(df_result)
-                print(df_result.shape)
-
+            r['source'](filepath)
+            
+            # get matrix
+            X = adata.X.todense().T
+            
+            def convert(X, func, gene_lengths=None):
+                # convert it into an R object to pass into the R function
+                with localconverter(robjects.default_converter + numpy2ri.converter):
+                    input_data_r = robjects.conversion.py2rpy(X)
+                # invoke the R function and get the result
+                if gene_lengths is None:
+                    result_r = func(input_data_r)
+                else:
+                    with localconverter(robjects.default_converter + numpy2ri.converter):
+                        gene_lengths_r = robjects.conversion.py2rpy(gene_lengths)
+                    result_r = func(input_data_r, gene_lengths_r)
+                # convert it back to a pandas dataframe (or numpy array)
+                with localconverter(robjects.default_converter + numpy2ri.converter):
+                    Y = robjects.conversion.rpy2py(result_r)
+                return Y
+                    
+            normalisation_methods = {
+                "TMM": robjects.globalenv['normalized_tmm_data'],
+                "DESeq": robjects.globalenv['normalized_deseq2_data'],
+                "TPM": robjects.globalenv['normalized_tpm_data'],
+            }
+            # check missing arguments for TPM
             if args.normalization_method == 'TPM':
-                # TPM normalization selected
-
-                # Check if length information is present
-                if 'length' in sampledata.columns:
-                    lengthdata = sampledata["length"]
-                else: 
-                    logger.critical("Column named \"length\" must be present and must contain gene lenght information when using TPM. ")
-                
-                # Remove length data from sample for clean sample data
-                sampledata = sampledata.drop('length', axis=1)
-
-                r['source'](filepath)
-                
-                # Load the function defined in R.
-                TPM = robjects.globalenv['normalized_tpm_data']
-
-                # Convert it into an R object to pass into the R function
-                with localconverter(robjects.default_converter + pandas2ri.converter):
-                    input_data_r = robjects.conversion.py2rpy(sampledata)
-                
-                with localconverter(robjects.default_converter + pandas2ri.converter):
-                    length_data_r = robjects.conversion.py2rpy(lengthdata)
-
-                #Invoke the R function and get the result
-                df_result_r = TPM(input_data_r, length_data_r)
-                
-                # Convert it back to a pandas dataframe.
-                with localconverter(robjects.default_converter + pandas2ri.converter):
-                    df_result = robjects.conversion.rpy2py(df_result_r)
-
-                logger.debug(df_result)
-                print(df_result)
-                print(df_result.shape)
-
-            else: 
-                # Unsupported normalization method selected
-                msg = f'An unsupported normalization method was selected'
-                logger.critical(msg)
+                try:
+                    gene_lengths = np.loadtxt(args.gene_lengths)
+                except Exception as e:
+                    msg = f"{e}\n Missing or cannot load [--gene-lengths] required " \
+                        "for TPM normalization. Terminating!"
+                    logger.critical(msg)
+                X_normalised = convert(X, 
+                                       normalisation_methods[args.normalization_method],
+                                       gene_lengths=gene_lengths)
+            else:
+                X_normalised = convert(X, 
+                                       normalisation_methods[args.normalization_method])
             
-            # Write the h5ad file
-            h5ad_dir = Path(args.dest, 'h5ad')
-            h5ad_dir.mkdir(parents=True, exist_ok=False)
-            
-            X = sc.sparse.csc_matrix(df_result) # mat is your matrix of counts normalized, etc.
-            adata = ad.AnnData(X.T)
-            adata.var_names = gene_metadata['name'].astype(str) # here features are e.g. gene names/symbols
-            adata.obs_names = sampledata.columns.astype(str)
-            
-            adata.var_names_make_unique()
-            adata.obs_names_make_unique()
-            
-            adata.var['gene_symbol'] = gene_metadata['name'].values.astype(str)
-            adata.var['gene_id'] = gene_metadata['ID'].values.astype(str)
+            X_normalised = csr_matrix(X_normalised.astype(np.float32)).T
+            adata.layers["raw"] = adata.X
+            adata.X = X_normalised
+    else:
+        msg = f"{config['dataset_type']} not supported!"
+        logger.critical(msg)
+        return
+        
+    # add missing features (id or symbol) to the final object
+    if kwargs['n_vars'] == 1:
+        qterm = "gene_id" if kwargs['feature'] == "gene_symbol" else "gene_symbol"
+        scopes = args.mygene_scopes if qterm == "gene_symbol" else "symbol"
+        adata = data_utils.add_missing_var_cols(adata,
+                                                qterm,
+                                                scopes,
+                                                args.mygene_scopes,
+                                                config['sample_taxid'],
+                                                kwargs['var_names']
+        )
 
-            # Sort the lists to create equal lists, if the same data is present
-            sampledata_headers_list = sampledata.columns.to_list()
-            observations_list = observations['title'].to_list()
-
-            sampledata_headers_list = sampledata_headers_list.sort() 
-            observations_list = observations_list.sort() 
-
-            if sampledata_headers_list != observations_list: 
-                logger.critical('count matrix columns do not match observations.tab.gz dataset. Make sure all of the data is present!')
-                print('count matrix columns do not match observations.tab.gz dataset. Make sure all of the data is present!')
-
-            # Observations.tab.gz
-            # Select the relevant information
-            # adata.obs can be populated from observations.tab.gz using minimally title, geo_accession, source_name_ch1, characteristics_ch1. 
-            # The field characteristics may contain multiple entries e.g. characteristics_ch1.1.cell type characteristics_ch1.2.age, etc.
-
-            # First get all the columns containing 'characteristics_ch1'
-            characteristics_ch1_cols: list[str] = [col for col in observations.columns if 'characteristics_ch1' in col]
-            headers_to_find = ['title','geo_accession','source_name_ch1'] + characteristics_ch1_cols
-            obs = observations[headers_to_find]
-            print(obs)
-
-            adata.obs_names = sampledata.columns # these are the sample names, they should match the column names from the count matrix, and the `title` from observations.tab.gz
-            adata.obs = obs # obs is a dataframe with observations from observations.tab.gz, where you selected the relevant entries
-            
-            adata.write_h5ad(Path(h5ad_dir, f'{args.name}_raw.h5ad'))
-            
-        else: 
-            # Data needs to be normalized, alert the user to the missing flag
-            msg = f'Data needs to be normalized when processing bulk data. Set the flag accordingly. '
-            logger.critical(msg)
-            return
+    # finally, write to disk
+    adata.write_h5ad(Path(h5ad_dir, h5ad_name))
+    msg = f"Writing data to {Path(h5ad_dir, h5ad_name).as_posix()}. Completed!"
+    logger.info(msg)
     
 
 if __name__ == '__main__':